Create Dating_Simulator
This program allows users to create a dating profile and the profiles of their potential dates on the simulator. Once the user has created as many profiles as they want, the program will randomly display them like on a real-life dating app. The user can then skip or match with the randomly generated profile. Once the user has matched with a profile, the program will randomly generate an ice-breaker to begin the conversation.